perf: float32 sweep for field values (preserve dtype, cast at source)
Central cfdmod/core/dtypes.py (FIELD_DTYPE=float32). The field-value ops now
PRESERVE their input field's dtype instead of upcasting to float64
(force/moment/statistics/field_series/extreme_value/derivative/moving_average/
frequency_filter/probe_extraction/profile_interpolation): float32 solver data
flows as float32 end to end (halving those arrays), while float64 sources stay
float64 so precision/exactness is unchanged. force_contribution /
moment_contribution infer the dtype from the field and cast the geometric
factors to match.

per_floor_loads casts its source pressure to FIELD_DTYPE (float32) so the whole
per-triangle load chain is float32 by default. Geometry, the structural/modal
solve, the time axis and the on-disk storage format stay float64.

chunk parity is now float32 (~1 ULP summation-order rounding across window
shapes), so that test uses a float32 tolerance; the face_cut exactness tests run
on the float64 fixture unchanged.

Refs #184

+++ cfdmod/core/ops/data_source_create/extreme_value.py
@@ -235,7 +235,7 @@ def extreme_value(ds: DataSource, p: ExtremeValueParams) -> DataSource:
             f"extreme_value requires at least 2 timesteps (got n_timesteps={ds.time.n_timesteps})"
         )
 
-    arr = np.asarray(ds.fields.read(p.field), dtype=np.float64)
+    arr = np.asarray(ds.fields.read(p.field))
     if arr.ndim != 2:
         raise ValueError(
             f"field {p.field!r} must be 2-D (n_elements, n_timesteps); got shape {arr.shape}"
